fix(api): tightening access control (#575)
* fix(api): tightening access control

* more review feedback

* test fixes

* pr feedback

* fix tests

* pr feedback

* add tests

* pr feedback

* fix tests

* fix classification config

* fixes

* more access control fixes

* pr cleanup

* more pr work

* more fixes

* ingestion changes

* updates to case service

* type fix

* fix collection error

* test fixes

* more fixes

* more test fixes

* yet more updates to case saving

* more fixes

* more test fixes

* more test fixes

* last fixes

* fix for sentinel tests

* added access control to delete endpoint

* imporve websocket access control

* small code fix

* fix tests

* more test fixes

* pr feedback

---------

+++ api/test/integration/api/test_wait_for_refresh.py
@@ -64,9 +64,9 @@ class SpyCollection:
         self._record_write("bulk", refresh)
         return self.wrapped_collection.bulk(operations, refresh)
 
-    def delete_by_query(self, query: str, sort=None, max_docs=None, refresh=None):
+    def delete_by_query(self, query: str, sort=None, max_docs=None, refresh=None, filters=None, access_control=None):
         self._record_write("delete_by_query", refresh)
-        return self.wrapped_collection.delete_by_query(query, sort, max_docs, refresh)
+        return self.wrapped_collection.delete_by_query(query, sort, max_docs, refresh, filters, access_control)
 
     def delete_by_search_object(self, query: dict, sort=None, max_docs=None, refresh=None):
         self._record_write("delete_by_search_object", refresh)
